Use map to describe routes in github lens
