+++ packages/github-lens-api/src/controller/describe.ts
@@ -1,42 +1,19 @@
 import { getDescribeRouterHandler } from 'common/expressRoutes';
 import type { Router } from 'express';
 
-function describeRoutes(path: string) {
-	let info = '';
-	switch (path) {
-		case '/healthcheck':
-			info = 'Display healthcheck';
-			break;
-		case '/repos':
-			info =
-				'Show all repos, with their team owners, optionally search by name with ?name=^repo-name-regex.*';
-			break;
-		case '/repos/:name':
-			info = 'Show repo and its team owners, if it exists';
-			break;
-		case '/archivedRepos':
-			info = 'Show all archived repos';
-			break;
-		case '/archivedReposNamesOnly':
-			info = 'Show all archived repos, names only (for tracker)';
-			break;
-		case '/teams':
-			info = 'Show all teams, with the repositories they own';
-			break;
-		case '/teams/:slug':
-			info = 'Show team and the repositories it owns, if it exists';
-			break;
-		case '/members':
-			info = 'Show member, with the teams they are in';
-			break;
-		case '/members/:login':
-			info = 'Show member and the teams they are in, if it exists';
-			break;
-		default:
-			info = 'No path info supplied';
-			break;
-	}
-	return info;
+function describeRoutes(path: string): string {
+	const infoByRoute: Record<string, string> = {
+		'/heathcheck': 'Display healthcheck',
+		'/repos':
+			'Show all repos, with their team owners, optionally search by name with ?name=^repo-name-regex.* and filter for ?isArchived',
+		'/repos/:name': 'Show repo and its team owners, if it exists',
+		'/teams': 'Show all teams, with the repositories they own',
+		'/teams/:slug': 'Show team with info, if it exists',
+		'/members': 'Show member, with the teams they are in',
+		'/members/:login': 'Show member and the teams they are in, if it exists',
+	};
+
+	return infoByRoute[path] ?? 'No path info supplied';
 }
 
 export const getRouteDescriptions = (router: Router) => {
